legalMind-rag/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "legalMind-rag"
version = "1.0.0"
description = "Production Naive RAG system for legal document Q&A"
requires-python = ">=3.11"
readme = "README.md"

dependencies = [
    # Web framework
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",

    # Async HTTP
    "httpx>=0.27.0",
    "aiohttp>=3.9.0",
    "aiofiles>=23.2.0",

    # OpenAI
    "openai>=1.30.0",

    # Vector store backends
    "chromadb>=0.5.0",
    "pinecone-client>=4.0.0",

    # Document loading
    "pdfplumber>=0.11.0",
    "python-docx>=1.1.0",
    "beautifulsoup4>=4.12.0",

    # Database
    "asyncpg>=0.29.0",
    "sqlalchemy[asyncio]>=2.0.0",

    # Cache
    "redis[hiredis]>=5.0.0",

    # Config & validation
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.0",

    # Auth
    "pyjwt>=2.8.0",

    # Logging
    "loguru>=0.7.0",

    # Observability
    "prometheus-client>=0.20.0",
    "opentelemetry-api>=1.24.0",
    "opentelemetry-sdk>=1.24.0",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",  # for AsyncClient in tests
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
]
reranker = [
    "sentence-transformers>=3.0.0",
]
eval = [
    "ragas>=0.1.10",
    "datasets>=2.19.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 99
target-version = "py311"
select = ["E", "W", "F", "I", "B", "UP"]
ignore = ["E501", "B008"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.hatch.build.targets.wheel]
packages = ["src"]
